+++ src/any_agent/evaluation/__init__.py
@@ -1,4 +1,5 @@
-from .evaluate import TraceEvaluationResult, evaluate
+from .evaluate import evaluate
 from .evaluation_case import EvaluationCase
+from .schemas import EvaluationResult, TraceEvaluationResult
 
-__all__ = ["EvaluationCase", "TraceEvaluationResult", "evaluate"]
+__all__ = ["EvaluationCase", "EvaluationResult", "TraceEvaluationResult", "evaluate"]
